+++ crates/db-storage/src/migrations/V28__tariffs_disabled_features.sql
@@ -1,0 +1,1 @@
+ALTER TABLE tariffs ADD COLUMN disabled_features TEXT[] DEFAULT ARRAY[]::TEXT[] NOT NULL;