+++ crates/signaling-core/src/module_tester.rs
@@ -617,6 +617,7 @@ where
                         name: "test".into(),
                         quotas: Default::default(),
                         enabled_modules: HashSet::from([M::NAMESPACE.into()]),
+                        disabled_features: Default::default(),
                     }
                     .into(),
                     module_data,