+++ crates/types/src/common.rs
@@ -5,5 +5,6 @@
 //! This module contains types that are used in different areas of the OpenTalk API,
 //! such as the Web API and signaling.
 
+pub mod features;
 pub mod shared_folder;
 pub mod tariff;